+++ Tests/Private/Test-TeamViewer32on64.Tests.ps1
@@ -1,0 +1,12 @@
+BeforeAll {
+    $Script:Module_RootPath = (Resolve-Path -Path (Join-Path -Path $PSScriptRoot -ChildPath '..\..'))
+    $Script:Module_PrivCmdletsPath = Join-Path -Path $Module_RootPath -ChildPath 'Cmdlets\Private'
+
+    . (Join-Path -Path $Module_PrivCmdletsPath -ChildPath 'Test-TeamViewer32on64.ps1')
+}
+
+Describe 'Test-TeamViewer32on64' {
+    It 'Should be available after dot-sourcing' {
+        Get-Command -Name 'Test-TeamViewer32on64' -CommandType Function -ErrorAction SilentlyContinue | Should -Not -BeNullOrEmpty
+    }
+}